Week3:structure changed as per the review comments
