+++ package.json
@@ -8,7 +8,7 @@
     "task1.2": "nodemon src/home_work1/task2.js",
     "task1.3": "babel-node src/home_work1/task3.js",
     "task2.1": "babel-node src/home_work2/task1.js",
-    "task3": "babel-node src/home_work3/task1.js"
+    "task3": "babel-node src/app.js"
   },
   "keywords": [
     "nodejs",
@@ -26,6 +26,8 @@
     "csv-parser": "^2.3.2",
     "csvtojson": "^2.0.10",
     "eslint": "^6.7.2",
+    "eslint-config-prettier": "^6.10.0",
+    "eslint-plugin-prettier": "^3.1.2",
     "event-stream": "^4.0.1",
     "express": "^4.17.1",
     "express-joi-validation": "^4.0.3",
@@ -33,12 +35,12 @@
     "http-status-codes": "^1.4.0",
     "joi": "^14.3.1",
     "mongoose": "^5.7.13",
-    "sequelize": "^5.21.3",
-    "stream": "0.0.2",
-    "uuid": "^3.3.3",
     "path": "^0.12.7",
     "pg": "^7.17.1",
-    "pg-hstore": "^2.3.3"
+    "pg-hstore": "^2.3.3",
+    "sequelize": "^5.21.3",
+    "stream": "0.0.2",
+    "uuid": "^3.3.3"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
