+++ src/models/UserModel.js
@@ -1,30 +1,10 @@
 const uuid = require("uuid/v4");
-let Sequelize = require("sequelize");
-let dbConnector = require("../db/dbConnector");
+const Sequelize = require("sequelize");
+const dbConnector = require("../middlewares/dbConnector");
+const users = require("./users")
 
-var User = dbConnector.define(
-  "users",
-  {
-    login: { type: Sequelize.STRING, allowNull: false },
-    password: {
-      type: Sequelize.STRING,
-      allowNull: false,
-      validate: {
-        validatePassword: function(value) {
-          if (!/^(?=.*[0-9])(?=.*[a-zA-Z])([a-zA-Z0-9]+)$/i.test(value)) {
-            throw new Error("Password does not meets the requirements");
-          }
-        }
-      }
-    },
-    age: {
-      type: Sequelize.INTEGER,
-      allowNull: false,
-      validate: { min: 4, max: 130 }
-    },
-    id: { type: Sequelize.UUID, primaryKey: true, defaultValue: Sequelize.UUIDV4 },
-    isDeleted: { type: Sequelize.BOOLEAN, defaultValue: false }
-  },
+const User = dbConnector.define(
+  "users",users,
   { freezeTableName: true }
 );
