+++ src/routers/userRoutes.js
@@ -1,23 +1,13 @@
-let express = require("express");
-let router = express.Router();
+const express = require("express");
+const router = express.Router();
 const Joi = require("joi");
 const validator = require("express-joi-validation").createValidator({});
 
-let userModel = require("../models/UserModel");
-let UserService = require("../services/userService");
-let userService = new UserService(userModel);
+const userModel = require("../models/UserModel");
+const UserService = require("../services/userService");
+const userService = new UserService(userModel);
 
-const bodySchema = Joi.object().keys({
-  login: Joi.string().required(),
-  password: Joi.string()
-    .regex(/^(?=.*[0-9])(?=.*[a-zA-Z])([a-zA-Z0-9]+)$/)
-    .required(),
-  age: Joi.number()
-    .integer()
-    .min(4)
-    .max(130)
-    .required()
-});
+const userSchema = require("../middlewares/validators").userSchema;
 
 router.get("/:id", (req, res) => {
   const id = req.params.id;
@@ -30,12 +20,12 @@ router.get("/:id", (req, res) => {
   });
 });
 
-router.post("/", validator.body(bodySchema), (req, res) => {
+router.post("/", validator.body(userSchema), (req, res) => {
   const body = req.body;
   userService.createUser(body).then(user => res.send(user));
 });
 
-router.put("/:id", validator.body(bodySchema), (req, res) => {
+router.put("/:id", validator.body(userSchema), (req, res) => {
   const body = req.body;
   const id = req.params.id;
   userService.updateUser(id, body).then(user => {
